furina/requirements.txt
py-cpuinfo
gputil
setuptools